Finished the game.
Changed everything in the kv file to root instead of self.parent where
possible.

Added the possibility to exit game to main menu. Hooked up the button.

Added the functionality of killing players, ending the game and victory.

Added Victory screen.

Added leaderboard loading and saving.

Changed wind, updated to match the mass change in previous commit.

Separated terrain generation into its own file.

Added shooting player to shell, to enable easier manipulation on hit
detection.

Changed size of the tank and gun barrel to be proportional to tank body,
which can be set as a constant in the main file.
In the future, maybe move it to be settable by the user.

Changed collision detection to return the player hit.

Fixed remove player to properly fix the current player index.